requirements.txt
flask
flask-login
flask-sqlalchemy
flask-celeryext
flask-wtf
celery
python-dotenv
psycopg2-binary
requests